Reorganized with <package>/<version> hierarchy
Instead of having a "flat" "<package>-<version>" hierarchy as before,
create a separate sub-directory for each package and a sub-sub-directory
for each version of that package (version, not conda build id!). This
helps us maintain builds for multiple versions of each package without
an messy explosion of subdirectories in the top level of the repository.