Recipe for BioBuilds 2015.11 release metapackage
Metapackage with the list of applications for the BioBuilds 2015.11
release; installing this package also generates a "pinned" file that
prevents unintentional upgrades/downgrades of applications and libraries
within the conda environment, reflecting the idea that each BioBuilds
release provides a single, point-in-time reference for a collection of
tools. Conda provides other tools for users wanting to create their own
reference collection of tools, so such pinning should not be an issue.