Make graphviz build identifier more specific (fixes #2)
Changed from simply using a number to "-bb0" for the build identifier.
This prevents users from accidentally installing the graphviz package
from conda's "defaults" channel (which as added since the 2015.11
release of BioBuilds); the "defaults" version has an external dependency
on libtool shared libraries (libltdl.so) that we'd like to avoid.