Pin "libgd" version in requirements (fixes #5)
A "libgd" package was added to the osx-64 "defaults" channel after the
2015.11 release. However, due to differences in the dylib install names,
the "biobuilds" and "defaults" packages are incompatible. Therefore,
we've modified the requirements for several packages to explicitly ask
for the BioBuilds "libgd" package.

WARNING: this fix is predicated on the assumption that Continuum is not
planning on adding packages for older libgd versions! We've had to leave
the build number off the requirements due to a conda bug that prevents
it from correctly and recursively installing libgd dependencies; see
<https://github.com/conda/conda/issues/2066> for details.