Recipe for scipy 0.17.0 (ppc64le only)
"scipy" is a dependency for some tools we want to add to BioBuilds, but
it's currently not available for ppc64le through the 'defaults' channel.

** WARNING **: releasing despite various tests failing; these failed
tests seem to mostly involve insufficient floating-point precision, so
fixing them will likely take quite a bit of time and involve working
with upstream. However, we don't want these issues to hold up other
BioBuilds development work, so we're just going to hope that they don't
adversely affect by the tools that depend on scipy.