Dockerfile & associated files for build image
Generate a Docker image that's essentially equivalent to the
"lab7io/build:biobuilds" image used by the "docker-build" script.

NOTE: this is *not* the actual Dockerfile used to generate the
"lab7io/build:biobuilds" image available on Docker Hub. That image was
built using a slightly different set of base images ("FROM" directives),
in order to reduce the number of container layers Lab7 Systems needs to
maintain internally. However, modulo some "FROM" and "MAINTAINER"
statements, the sequence of commands here is identical to the sequence
of commands Lab7 Systems used to produce the Docker Hub image.