Switch to 'bioconductor-*' package naming convention
Follow the bioconda project's lead and use 'bioconductor-*' to identify
BioConductor R packages, instead of simply using the 'r-*' convention
and confusing them with packages available from CRAN. Adjusted the
"requirements" sections of affected packages to reflect this change, and
gave them a build number bump for good measure.

NOTE: we did *NOT* alter the dependency names in already-released
"biobuilds" packages, as doing that would break our commitment to keep
them as a "reference" collection of binaries. Basically, we'll just have
to keep those not-so-brilliantly-named packages around indefinitely.