Recipes for other "missing" linux-ppc64le packages
Like matplotlib in the previous commit (4c25db2), these packages are
provided through Continuum's "defaults" channel for the x86_64 platforms
we support (i.e., "osx-64" and "linux-64") but not for little-endian
POWER (linux-ppc64le). The choice of packages is based on what's needed
to build "scikit-bio", and their versions are based on what's available
on the linux-64 "defaults" channel as of 2016-11-20.