Docker build images: require requests <=2.11 in conda root
Installing requests 2.12.* in the conda root environment severely breaks
conda and conda-build, and leaves us unable to use the `docker-build`
script to build BioBuilds packages. It's not clear if future versions of
requests and/or conda will fix this issue, but since the BioBuilds
Docker build images are meant to be "archival", we'll modify the "conda
install" command and "conda-meta/pinned" file to make sure we don't
install of requests 2.12 or later.

This bug has already been reported to the upstream (conda) developers;
see <https://github.com/conda/conda/issues/3928> for details.

(Lab7 ticket: BIOBUILDS-261)