Docker build images: fixed check for existing package dirs
The "conda-build.sh" entry point scripts were missing "-d" tests when
checking for the directories containing previously built packages. This
caused them to always add the local "conda-bld" directory as a conda
channel, which is not a problem *until* you try to build in a
freshly-cloned repository (i.e., one that hasn't been previously been
used for building BioBuilds packages).