Update links for Miniconda installers in release notes
Download links for the "latest" Miniconda installer packages (i.e.,
"Miniconda-latest-${os}-${arch}.sh") seem to be less reliable for
certain architectures (particularly POWER and ARMv7), so we've modified
the installer links to point to specific, versioned packages. This also
helps with our "better reproducibility" pitch as it's now clearer which
Miniconda installer was the "latest" at the time of each release.