libgd now explicitly depends on (lib)jpeg 8d
Version 9b is now available in the "defaults" channel, so we need to
make sure libgd uses the correct version of libjpeg (i.e., the one that
was used to build it).

NOTE: the "libgd" packages in the "biobuilds" channel were _not_
actually rebuilt (i.e., the binaries are identical to the ones in the
"-0" packages). We simply modified the appropriate files in the "info"
directory and used "tar" to create the new ("-1") packages.