Rebuild libgtextutils, fastx-toolkit on linux-64 with gcc 4.8
Resolves #12 (BIOBUILDS-339), wherein tools like fastq_quality_filter
were consistently segfaulting when running on x86_64 Linux. Expanded the
fastx-toolkit test suite to run the tools on actual data, rather than
just invoking them with the "-h" argument (which doesn't seem sufficient
to trigger the observed segfaults).

[Note: still unclear why the original gcc4.4-compiled versions failed,
but since this fix works, I'm not going to worry too much about it.]