`docker-build`: add `--no-pkg-cache` option
This option stops the `conda-bld/pkg_cahce` directory from being passed
as a channel ("-c") argument to `conda build` in the build container.
This, in effect, forces the build system to go out and fetch the latest
packages from the "defaults" channel, rather than using the cached
versions of these packages we happen to have on our build system.