Recipes to update cairo and its dependencies
Updating to R 3.4.2 so we might as well update its dependencies as well.
These packages are already available for x86_64 Linux and macOS, but
we're rebuilding them so variants with conda-build 3.0+ features (e.g.,
version pinning) are available for Linux on POWER (linux-ppc64le). Their
recipes are based on the recipes for the following linux-64 packages
from the "defaults" channel:

  - libxml2-2.9.4-h6b072ca_5
  - gperf-3.1-hb29a3a9_0
  - pcre-8.41-hc71a17e_0
  - gettext-0.19.8.1-hb0f4f8b_2 (from osx-64 "defaults")
  - glib-2.53.6-h5d9569c_2
  - pixman-0.34.0-hceecf20_3
  - freetype-2.8-h52ed37b_0
  - fontconfig-2.12.4-h88586e7_1
  - cairo-1.14.10-haa5651f_5

(Lab7 ticket: BIOBUILDS-342)