Recipes to update pango and its dependencies
Updating to R 3.4.2 so we might as well update its dependencies as well. These
packages are already available for x86_64 Linux and macOS, but we're rebuilding
them so variants with conda-build 3.0+ features (e.g., version pinning) are
available for Linux on POWER (linux-ppc64le). Their recipes are based on the
recipes for the following linux-64 packages from the "defaults" channel:

  - gobject-introspection-1.53.5-py27h26494a4
  - graphite2-1.3.10-hc526e54_0
  - harfbuzz-1.5.0-h2545bd6_0
  - pango-1.40.11-h8191d47_0

(Lab7 ticket: BIOBUILDS-342)