Recipe for OpenBLAS 0.2.20
Added mostly so we have an updated version available for linux-ppc64le.
Building for all platforms because the packaging structure does not
match that of the "defaults" channel (i.e., we don't split it up into
"openblas-devel", "libopenblas", and "libopenblas-static" sub-packages),
and we want to minimize the chances of that difference breaking
dependencies from other BioBuilds packages.

(Lab7 ticket: BIOBUILDS-345)