biobuilds-build: set `-mtune=haswell` for x86_64 targets
Copying option from Anaconda's default set of flags. We continue to
limit generated code to the Nehalem microarchitecture but now allow GCC
to arrange things for better performance on newer microarchitectures.
(We could have used `-mtune=intel` for a similar effect, but that option
may be particularly sensitive to the version of GCC we are running.)